saarthi_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email-validator]>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
# Auth
PyJWT>=2.8.0
bcrypt>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
